+++ Development/nmos-cpp-node/node_resources_thread.h
